Add files via upload
Project Overview:

Homepage Redesign: The primary focus of this project was to enhance the visual appeal and user experience of the restaurant's homepage. I applied modern design principles and optimized the layout for both desktop and mobile devices to ensure a responsive design.

Login Page Integration: To enhance user engagement, I integrated a secure login page accessible through a prominent login button on the homepage. This allows restaurant customers to access exclusive offers, make reservations, and leave reviews with ease.